refactor(update): split update-command into focused modules under the line limit
