+++ src/cli/update-cli/update-command-git.ts
@@ -1,0 +1,219 @@
+import type { UpdateChannel } from "../../infra/update-channels.js";
+import {
+  createGlobalInstallEnv,
+  globalInstallArgs,
+  resolveGlobalInstallTarget,
+  resolvePnpmGlobalDirFromGlobalRoot,
+} from "../../infra/update-global.js";
+import { runGatewayUpdate, type UpdateRunResult } from "../../infra/update-runner.js";
+import { defaultRuntime } from "../../runtime.js";
+import {
+  OPENCLAW_DATABASE_SCHEMA_DOCS_URL,
+  preflightOpenClawDatabaseSchemas,
+  type IncompatibleOpenClawDatabase,
+  type IndeterminateOpenClawDatabase,
+  type OpenClawDatabaseSchemaPreflight,
+} from "../../state/openclaw-database-preflight.js";
+import type { OpenClawSchemaVersions } from "../../state/openclaw-schema-versions.js";
+import { createUpdateProgress, printResult } from "./progress.js";
+import {
+  createGlobalCommandRunner,
+  ensureGitCheckout,
+  resolveGitInstallDir,
+  resolveGlobalManager,
+  runUpdateStep,
+  type UpdateCommandOptions,
+} from "./shared.js";
+import { UpdateCommandAbort, type PreManagedServiceStop } from "./update-command-service.js";
+
+const DEFAULT_UPDATE_STEP_TIMEOUT_MS = 30 * 60_000;
+
+type BeforeGitMutation = (target: {
+  schemaVersions?: OpenClawSchemaVersions;
+  metadataUnreadable?: string;
+}) => Promise<{
+  allowGatewayServiceRepair?: boolean;
+  allowGatewayActivation?: boolean;
+} | void>;
+
+export function formatSchemaRefusalLines(
+  schemas: {
+    incompatible: readonly IncompatibleOpenClawDatabase[];
+    indeterminate: readonly IndeterminateOpenClawDatabase[];
+  },
+  dryRun = false,
+): string[] {
+  const prefix = dryRun ? "Would refuse update" : "Update refused";
+  return [
+    ...schemas.incompatible.map((database) => {
+      const agent = database.agentId ? ` (agent ${database.agentId})` : "";
+      return `${prefix}: ${database.kind} database${agent} ${database.path} has schema ${database.foundVersion}; target supports ${database.supportedVersion}; writer build ${database.writerAppVersion ?? "unknown"}.`;
+    }),
+    ...schemas.indeterminate.map(
+      (database) =>
+        `${prefix}: could not inspect ${database.kind} database ${database.path}: ${database.reason}; retry once the gateway releases it.`,
+    ),
+    OPENCLAW_DATABASE_SCHEMA_DOCS_URL,
+    "Installing manually via npm bypasses this guard; back up first and verify compatibility.",
+  ];
+}
+
+export function checkTargetDatabaseSchemas(
+  supportedVersions: OpenClawSchemaVersions | undefined,
+  env: NodeJS.ProcessEnv = process.env,
+): OpenClawDatabaseSchemaPreflight {
+  return supportedVersions
+    ? preflightOpenClawDatabaseSchemas({ env, supportedVersions })
+    : { incompatible: [], indeterminate: [] };
+}
+
+export function hasSchemaRefusal(schemas: OpenClawDatabaseSchemaPreflight): boolean {
+  return schemas.incompatible.length > 0 || schemas.indeterminate.length > 0;
+}
+
+export function createBeforeGitMutation(params: {
+  roots: readonly string[];
+  shouldRestart: boolean;
+  stopManagedService: (roots: readonly string[]) => Promise<void>;
+  getPreManagedServiceStop: () => PreManagedServiceStop | undefined;
+  markSchemaRefusalAfterStop: () => void;
+}): BeforeGitMutation {
+  return async (target) => {
+    if (target?.metadataUnreadable) {
+      defaultRuntime.error(
+        `Update refused: could not inspect the target's schema support (${target.metadataUnreadable}). Retry, or see ${OPENCLAW_DATABASE_SCHEMA_DOCS_URL}.`,
+      );
+      defaultRuntime.exit(1);
+      throw new UpdateCommandAbort();
+    }
+    const preStopSchemas = checkTargetDatabaseSchemas(target?.schemaVersions);
+    if (hasSchemaRefusal(preStopSchemas)) {
+      defaultRuntime.error(formatSchemaRefusalLines(preStopSchemas).join("\n"));
+      defaultRuntime.exit(1);
+      throw new UpdateCommandAbort();
+    }
+    await params.stopManagedService(params.roots);
+    const preManagedServiceStop = params.getPreManagedServiceStop();
+    const postStopSchemas = checkTargetDatabaseSchemas(
+      target?.schemaVersions,
+      preManagedServiceStop?.serviceEnv ?? process.env,
+    );
+    if (hasSchemaRefusal(postStopSchemas)) {
+      params.markSchemaRefusalAfterStop();
+      defaultRuntime.error(formatSchemaRefusalLines(postStopSchemas).join("\n"));
+      throw new UpdateCommandAbort();
+    }
+    return {
+      // Only a positively owned service may be rewritten. Activation
+      // additionally requires this update to have stopped it.
+      allowGatewayServiceRepair: preManagedServiceStop?.serviceMatchesMutationRoot === true,
+      allowGatewayActivation:
+        params.shouldRestart &&
+        preManagedServiceStop?.stopped === true &&
+        preManagedServiceStop.serviceMatchesMutationRoot === true,
+    };
+  };
+}
+
+export async function runGitUpdate(params: {
+  root: string;
+  switchToGit: boolean;
+  installKind: "git" | "package" | "unknown";
+  timeoutMs: number | undefined;
+  startedAt: number;
+  progress: ReturnType<typeof createUpdateProgress>["progress"];
+  channel: UpdateChannel;
+  tag: string;
+  showProgress: boolean;
+  opts: UpdateCommandOptions;
+  stop: () => void;
+  devTargetRef?: string;
+  beforeGitMutation?: BeforeGitMutation;
+  allowGatewayServiceRepair: boolean;
+  allowGatewayActivation: boolean;
+}): Promise<UpdateRunResult> {
+  const updateRoot = params.switchToGit ? resolveGitInstallDir() : params.root;
+  const effectiveTimeout = params.timeoutMs ?? DEFAULT_UPDATE_STEP_TIMEOUT_MS;
+  const installEnv = await createGlobalInstallEnv();
+
+  const cloneStep = params.switchToGit
+    ? await ensureGitCheckout({
+        dir: updateRoot,
+        env: installEnv,
+        timeoutMs: effectiveTimeout,
+        progress: params.progress,
+      })
+    : null;
+
+  if (cloneStep && cloneStep.exitCode !== 0) {
+    const result: UpdateRunResult = {
+      status: "error",
+      mode: "git",
+      root: updateRoot,
+      reason: cloneStep.name,
+      steps: [cloneStep],
+      durationMs: Date.now() - params.startedAt,
+    };
+    params.stop();
+    printResult(result, { ...params.opts, hideSteps: params.showProgress });
+    defaultRuntime.exit(1);
+    return result;
+  }
+
+  const updateResult = await runGatewayUpdate({
+    cwd: updateRoot,
+    argv1: params.switchToGit ? undefined : process.argv[1],
+    timeoutMs: params.timeoutMs,
+    progress: params.progress,
+    channel: params.channel,
+    tag: params.tag,
+    devTargetRef: params.devTargetRef,
+    deferConfiguredPluginInstallRepair: true,
+    allowGatewayServiceRepair: params.allowGatewayServiceRepair,
+    allowGatewayActivation: params.allowGatewayActivation,
+    beforeGitMutation: params.beforeGitMutation,
+  });
+  const steps = [...(cloneStep ? [cloneStep] : []), ...updateResult.steps];
+
+  if (params.switchToGit && updateResult.status === "ok") {
+    const manager = await resolveGlobalManager({
+      root: params.root,
+      installKind: params.installKind,
+      timeoutMs: effectiveTimeout,
+    });
+    const runCommand = createGlobalCommandRunner();
+    const installTarget = await resolveGlobalInstallTarget({
+      manager,
+      runCommand,
+      timeoutMs: effectiveTimeout,
+      pkgRoot: params.root,
+    });
+    const installLocation =
+      installTarget.manager === "pnpm"
+        ? resolvePnpmGlobalDirFromGlobalRoot(installTarget.globalRoot)
+        : null;
+    const installStep = await runUpdateStep({
+      name: "global install",
+      argv: globalInstallArgs(installTarget, updateRoot, undefined, installLocation, updateRoot),
+      cwd: updateRoot,
+      env: installEnv,
+      timeoutMs: effectiveTimeout,
+      progress: params.progress,
+    });
+    steps.push(installStep);
+
+    const failedStep = installStep.exitCode !== 0 ? installStep : null;
+    return {
+      ...updateResult,
+      status: updateResult.status === "ok" && !failedStep ? "ok" : "error",
+      steps,
+      durationMs: Date.now() - params.startedAt,
+    };
+  }
+
+  return {
+    ...updateResult,
+    steps,
+    durationMs: Date.now() - params.startedAt,
+  };
+}
